pyproject.toml
[project]
name = "simple-story-llm"
version = "1.0.0"
description = "-"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama",
    "openai",
    "pyyaml",
    "google-generativeai",
]
